Fix #843

+++ gedbrowserng/src/main/java/org/schoellerfamily/gedbrowser/api/controller/PersonsController.java
@@ -2,11 +2,21 @@ package org.schoellerfamily.gedbrowser.api.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.schoellerfamily.gedbrowser.analytics.LivingEstimator;
+import org.schoellerfamily.gedbrowser.analytics.calendar.CalendarProvider;
 import org.schoellerfamily.gedbrowser.api.crud.ObjectCrud;
 import org.schoellerfamily.gedbrowser.api.crud.PersonCrud;
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiPerson;
+import org.schoellerfamily.gedbrowser.api.datamodel.ApiPerson.Builder;
+import org.schoellerfamily.gedbrowser.datamodel.Person;
+import org.schoellerfamily.gedbrowser.security.service.UserService;
+import org.schoellerfamily.gedbrowser.security.util.RequestUserUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -27,6 +37,14 @@ public class PersonsController extends CrudInvoker {
     /** Logger. */
     private final transient Log logger = LogFactory.getLog(getClass());
 
+    /** */
+    @Autowired
+    private UserService userService;
+
+    /** */
+    @Autowired
+    private transient CalendarProvider provider;
+
     /**
      * @return the CRUD object for manipulating persons
      */
@@ -35,14 +53,21 @@ public class PersonsController extends CrudInvoker {
     }
 
     /**
+     * @param request the servlet request object
      * @param db the name of the db to access
      * @param person the data for the person
      * @return the person as created
      */
     @PostMapping(value = "/v1/dbs/{db}/persons")
     @ResponseBody
-    public ApiPerson create(@PathVariable final String db,
+    public ApiPerson create(
+            final HttpServletRequest request,
+            @PathVariable final String db,
             @RequestBody final ApiPerson person) {
+        final RequestUserUtil requestUserUtil = new RequestUserUtil(request, userService);
+        if (!requestUserUtil.hasAdmin()) {
+            throw new AccessDeniedException("go away");
+        }
         return crud().createOne(db, person);
     }
 
@@ -58,6 +83,7 @@ public class PersonsController extends CrudInvoker {
     }
 
     /**
+     * @param request the servlet request coming in
      * @param db the name of the db to access
      * @param id the ID of the person
      * @return the person
@@ -65,13 +91,43 @@ public class PersonsController extends CrudInvoker {
     @GetMapping(value = "/v1/dbs/{db}/persons/{id}")
     @ResponseBody
     public ApiPerson read(
+            final HttpServletRequest request,
             @PathVariable final String db,
             @PathVariable final String id) {
+        final Person person = ((PersonCrud) crud()).read(db, id).getGedObject();
+        if (shouldHideLiving(request, person)) {
+            return createDummyLivingPerson(id);
+        }
         logger.info("entering read person: " + id);
         return crud().readOne(db, id);
     }
 
     /**
+     * Determine whether to return dummy "living" person.
+     *
+     * @param request the request, used to identify user authorization
+     * @param person the person, used to check if living
+     * @return true if the person is living and should be hidden
+     */
+    private boolean shouldHideLiving(final HttpServletRequest request, final Person person) {
+        return !new RequestUserUtil(request, userService).hasUser()
+                && new LivingEstimator(person, provider).estimate();
+    }
+
+    /**
+     * Create a minimal person for return, who is only identified as living.
+     *
+     * @param id the person ID
+     * @return the dummy person
+     */
+    private ApiPerson createDummyLivingPerson(final String id) {
+        final Builder builder =
+                new ApiPerson.Builder().id(id).indexName("Living").surname("").build();
+        return new ApiPerson(builder);
+    }
+
+    /**
+     * @param request the servlet request coming in
      * @param db the name of the db to access
      * @param id the id of the person to update
      * @param person the data for the person
@@ -79,14 +135,21 @@ public class PersonsController extends CrudInvoker {
      */
     @PutMapping(value = "/v1/dbs/{db}/persons/{id}")
     @ResponseBody
-    public ApiPerson update(@PathVariable final String db,
+    public ApiPerson update(
+            final HttpServletRequest request,
+            @PathVariable final String db,
             @PathVariable final String id,
             @RequestBody final ApiPerson person) {
+        final RequestUserUtil requestUserUtil = new RequestUserUtil(request, userService);
+        if (!requestUserUtil.hasAdmin()) {
+            throw new AccessDeniedException("go away");
+        }
         logger.info("entering update person: " + id);
         return crud().updateOne(db, id, person);
     }
 
     /**
+     * @param request the servlet request coming in
      * @param db the name of the db to access
      * @param id the ID of the person
      * @return the deleted person object
@@ -94,8 +157,13 @@ public class PersonsController extends CrudInvoker {
     @DeleteMapping(value = "/v1/dbs/{db}/persons/{id}")
     @ResponseBody
     public ApiPerson delete(
+            final HttpServletRequest request,
             @PathVariable final String db,
             @PathVariable final String id) {
+        final RequestUserUtil requestUserUtil = new RequestUserUtil(request, userService);
+        if (!requestUserUtil.hasAdmin()) {
+            throw new AccessDeniedException("go away");
+        }
         logger.info("entering delete person: " + id);
         return crud().deleteOne(db, id);
     }
